refactor: move process tables to a utility
